Move all NotFoundExceptions to notfound package
