+++ src/main/java/com/cvgenerator/exceptions/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.cvgenerator.exceptions;
 
+import com.cvgenerator.exceptions.notfound.*;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
