+++ src/main/java/com/cvgenerator/exceptions/notfound/UserCvNotFoundException.java
@@ -1,4 +1,4 @@
-package com.cvgenerator.exceptions;
+package com.cvgenerator.exceptions.notfound;
 
 import lombok.extern.slf4j.Slf4j;
