+++ src/main/java/com/cvgenerator/service/implementation/CourseServiceImpl.java
@@ -3,8 +3,8 @@ package com.cvgenerator.service.implementation;
 import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Course;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.CourseNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.CourseNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.CourseRepository;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.service.CourseService;
