+++ src/main/java/com/cvgenerator/service/implementation/EducationServiceImpl.java
@@ -4,7 +4,7 @@ import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Education;
 import com.cvgenerator.domain.entity.UserCv;
 import com.cvgenerator.exceptions.EducationNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.EducationRepository;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.service.EducationService;
