+++ src/main/java/com/cvgenerator/service/implementation/ExperienceServiceImpl.java
@@ -3,8 +3,8 @@ package com.cvgenerator.service.implementation;
 import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Experience;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.ExperienceNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.ExperienceNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.ExperienceRepository;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.service.ExperienceService;
