+++ src/main/java/com/cvgenerator/service/implementation/InterestServiceImpl.java
@@ -3,8 +3,8 @@ package com.cvgenerator.service.implementation;
 import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Interest;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.InterestNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.InterestNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.InterestRepository;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.service.InterestService;
