+++ src/main/java/com/cvgenerator/service/implementation/PersonalDataServiceImpl.java
@@ -4,9 +4,8 @@ import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Address;
 import com.cvgenerator.domain.entity.PersonalData;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.PersonalDataNotFoundException;
-import com.cvgenerator.exceptions.ProjectNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.PersonalDataNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.AddressRepository;
 import com.cvgenerator.repository.PersonalDataRepository;
 import com.cvgenerator.repository.UserCvRepository;
