+++ src/main/java/com/cvgenerator/service/implementation/ProjectServiceImpl.java
@@ -3,8 +3,8 @@ package com.cvgenerator.service.implementation;
 import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Project;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.ProjectNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.ProjectNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.ProjectRepository;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.service.ProjectService;
