+++ src/main/java/com/cvgenerator/service/implementation/SkillServiceImpl.java
@@ -3,8 +3,8 @@ package com.cvgenerator.service.implementation;
 import com.cvgenerator.config.Messages;
 import com.cvgenerator.domain.entity.Skill;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.SkillNotFoundException;
-import com.cvgenerator.exceptions.UserCvNotFoundException;
+import com.cvgenerator.exceptions.notfound.SkillNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserCvNotFoundException;
 import com.cvgenerator.repository.SkillRepository;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.service.SkillService;
