+++ src/main/java/com/cvgenerator/service/implementation/UserServiceImpl.java
@@ -5,7 +5,7 @@ import com.cvgenerator.domain.dto.UserCvShortDto;
 import com.cvgenerator.domain.dto.UserDto;
 import com.cvgenerator.domain.entity.User;
 import com.cvgenerator.domain.entity.UserCv;
-import com.cvgenerator.exceptions.UserNotFoundException;
+import com.cvgenerator.exceptions.notfound.UserNotFoundException;
 import com.cvgenerator.repository.UserCvRepository;
 import com.cvgenerator.repository.UserRepository;
 import com.cvgenerator.service.UserService;
@@ -13,7 +13,6 @@ import com.cvgenerator.service.dtoConverters.UserCvShortDtoConverter;
 import com.cvgenerator.service.dtoConverters.UserDtoConverter;
 import com.cvgenerator.utils.service.implementation.MailServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.MessageSource;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
